pyproject.toml
[project]
name = "youtube-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "012MykA",email = "1deadeye1@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "cryptography (>=46.0.3,<47.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
